+++ src/compas_plotters/plotter.py
@@ -264,9 +264,9 @@ class Plotter(metaclass=Singleton):
         if not artist:
             if self.zstack == 'natural':
                 zorder = 1000 + len(self._artists) * 100
-                artist = PlotterArtist(item, zorder=zorder, **kwargs)
+                artist = PlotterArtist(item, zorder=zorder, context='Plotter', **kwargs)
             else:
-                artist = PlotterArtist(item, **kwargs)
+                artist = PlotterArtist(item, context='Plotter', **kwargs)
         artist.draw()
         self._artists.append(artist)
         return artist
@@ -283,7 +283,7 @@ class Plotter(metaclass=Singleton):
                artist_type: PlotterArtist,
                **kwargs) -> PlotterArtist:
         """Add a COMPAS geometry object or data structure using a specific artist type."""
-        artist = PlotterArtist(item, artist_type=artist_type, **kwargs)
+        artist = PlotterArtist(item, artist_type=artist_type, context='Plotter', **kwargs)
         artist.draw()
         self._artists.append(artist)
         return artist